% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/support.R
=======
% Please edit documentation in R/general.R
>>>>>>> 38802c698b7513bb5bd529d7c210b18f5081a4d5
\name{add_variable_labels}
\alias{add_variable_labels}
\title{support function to plot variable text labels}
\usage{
add_variable_labels(p, ...)
}
\arguments{
\item{p}{plot made with jointVIP object}

\item{...}{encompasses other variables needed}
}
\value{
a joint variable importance plot of class \code{ggplot} with curves
}
\description{
support function to plot variable text labels
}
