% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{bootstrap.plot}
\alias{bootstrap.plot}
\title{plot the bootstrap version of the jointVIP object}
\usage{
bootstrap.plot(
  x,
  ...,
  smd = "OVB-based",
  use_abs = TRUE,
  plot_title = "Joint Variable Importance Plot",
  B = 100
)
}
\arguments{
\item{x}{a jointVIP object}

\item{...}{custom options: \code{bias_curve_cutoffs}, \code{text_size}, \code{max.overlaps}, \code{label_cut_std_md}, \code{label_cut_outcome_cor}, \code{label_cut_bias}, \code{OVB_curves}, \code{add_var_labs}}

\item{smd}{specify the standardized mean difference is \code{OVB-based} or \code{standard}}

\item{use_abs}{TRUE (default) for absolute measures}

\item{plot_title}{optional string for plot title}

\item{B}{100 (default) for the number of times the bootstrap step wished to run}
}
\value{
a joint variable importance plot of class \code{ggplot}
}
\description{
plot the bootstrap version of the jointVIP object
}
