% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_boot_variation.R
\name{calc_bootstrap}
\alias{calc_bootstrap}
\title{Using bootstap to calculate the confidence bars}
\usage{
calc_bootstrap(
  B,
  pilot_df,
  analysis_df,
  covariates,
  treatment,
  outcome,
  use_denom,
  use_abs = F
)
}
\arguments{
\item{B}{number of bootstrap to do}

\item{pilot_df}{dataframe of the pilot data}

\item{analysis_df}{dataframe of the analysis data}

\item{covariates}{vector of strings or list denoting column names of interest}

\item{treatment}{string denoting the name of the treatment variable}

\item{outcome}{string denoting the name of the outcome variable}

\item{use_denom}{which denom to use}

\item{use_abs}{boolean to denote whether absolute values are used to construct the measures}
}
\description{
Using bootstap to calculate the confidence bars
}
