% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/check_measures.R
=======
% Please edit documentation in R/measures.R
>>>>>>> 38802c698b7513bb5bd529d7c210b18f5081a4d5
\name{check_measures}
\alias{check_measures}
\title{Check measures
Check to see if there is any missing values or variables
without any variation or identical rows (only unique rows will be used)}
\usage{
check_measures(measures)
}
\arguments{
\item{measures}{measures needed for jointVIP}
}
\value{
measures needed for jointVIP
}
\description{
Check measures
Check to see if there is any missing values or variables
without any variation or identical rows (only unique rows will be used)
}
