% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{create_post_jointVIP}
\alias{create_post_jointVIP}
\title{create jointVIP object}
\usage{
create_post_jointVIP(object, post_analysis_df)
}
\arguments{
\item{object}{a jointVIP object}

\item{post_analysis_df}{post matched or weighted data.frame}
}
\value{
a post_jointVIP object (subclass of jointVIP)
}
\description{
This is creates the post_jointVIP object & check inputs
}
\examples{
# not run
# new_jointVIP <- create_jointVIP(treatment,
#                                 outcome,
#                                 covariates,
#                                 pilot_df,
#                                 analysis_df)
## at this step typically you may wish to do matching or weighting
## the results after can be stored as a post_data
## the post_data here is not matched or weighted, only for illustrative purposes
# post_data <- data.frame(year = rnorm(50, 200, 5),
#                         pop = rnorm(50, 1000, 500),
#                         gdpPercap = runif(50, 100, 1000),
#                         trt = rbinom(50, 1, 0.5),
#                         out = rnorm(50, 1, 0.2))
# create_post_jointVIP(new_jointVIP, post_data)
}
