% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_measure.R
\name{get_analysis_denom}
\alias{get_analysis_denom}
\title{Get standardized difference in analysis sample}
\usage{
get_analysis_denom(covariate, analysis_sample, treatment)
}
\arguments{
\item{covariate}{covariate, note it is each column of the sample}

\item{analysis_sample}{sample for analysis containing treatment and control}

\item{treatment}{string denoting the name of the treatment variable}
}
\description{
Get standardized difference in analysis sample
}
