% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in R/get_boot_measures.R
=======
% Please edit documentation in R/measures.R
>>>>>>> 38802c698b7513bb5bd529d7c210b18f5081a4d5
=======
% Please edit documentation in R/measures.R
>>>>>>> 38802c698b7513bb5bd529d7c210b18f5081a4d5
\name{get_boot_measures}
\alias{get_boot_measures}
\title{Calculate bootstrapped variation
additional tool to help calculate the uncertainty of each variable's bias}
\usage{
get_boot_measures(object, smd = "cross-sample", use_abs = TRUE, B = 100)
}
\arguments{
\item{object}{jointVIP object}

\item{smd}{calculate standardized mean difference either using \code{cross-sample} or \code{pooled}}

\item{use_abs}{TRUE (default) for absolute measures}

\item{B}{100 (default) for the number of times the bootstrap step wished to run}
}
\value{
bootstrapped measures needed for bootstrap-jointVIP
}
\description{
Calculate bootstrapped variation
additional tool to help calculate the uncertainty of each variable's bias
}
