% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jointVIP.R
\name{get_jointVIP}
\alias{get_jointVIP}
\title{Get the measures and create plot}
\usage{
get_jointVIP(
  measures,
  bias_curve_cutoffs = c(-0.05, -0.03, -0.01, -0.005, 0.005, 0.01, 0.03, 0.05),
  use_abs = F,
  use_denom = "both",
  plot_title = "Joint variable importance",
  pre_matched_alpha = 1,
  label_cutoff_std_diff = NULL,
  label_cutoff_control_cor = NULL,
  label_cutoff_bias = NULL,
  raw_plot = F,
  max.overlapss = 10,
  point_text_size = 3
)
}
\arguments{
\item{measures}{the measures calculated for jointVIP}

\item{bias_curve_cutoffs}{cut offs of the bias curve one wishes to plot}

\item{use_abs}{boolean to denote whether absolute values are used to construct the measures}

\item{use_denom}{must be "standard", "pilot", or "min"}

\item{plot_title}{title of the plot to plot}

\item{pre_matched_alpha}{set to 1; changed if post-match exists}

\item{label_cutoff_std_diff}{text label cut off for standardized difference}

\item{label_cutoff_control_cor}{text label cut off for correlation}

\item{label_cutoff_bias}{text label cut off for bias}

\item{raw_plot}{plot without bias}

\item{max.overlapss}{maximum overlap of points}

\item{point_text_size}{text size for each point}
}
\description{
Get the measures and create plot
}
