% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_measure.R
\name{get_pilot_cor}
\alias{get_pilot_cor}
\title{Get correlation in pilot sample}
\usage{
get_pilot_cor(covariate, pilot_sample, outcome)
}
\arguments{
\item{covariate}{covariate, note it is each column of the sample}

\item{pilot_sample}{sample for analysis containing only control}

\item{outcome}{string denoting the name of the outcome variable}
}
\description{
Get correlation in pilot sample
}
