% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_measures.R
\name{get_pilot_denom}
\alias{get_pilot_denom}
\title{Get denominator using the pilot sample
this is done by getting the standard deviation of the variable.}
\usage{
get_pilot_denom(covariate)
}
\arguments{
\item{covariate}{covariate, note it is each column of the sample}
}
\description{
Get denominator using the pilot sample
this is done by getting the standard deviation of the variable.
}
