% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_post_measures.R
\name{get_post_analysis_vip}
\alias{get_post_analysis_vip}
\title{Get standardized difference in analysis sample}
\usage{
get_post_analysis_vip(
  post_analysis_df,
  measures,
  props,
  progs,
  treatment,
  outcome,
  covariates,
  use_denom,
  joint_vip,
  use_abs,
  post_prop = NULL,
  post_prog = NULL
)
}
\arguments{
\item{post_analysis_df}{after matching data frame}

\item{measures}{measures from calculate measures}

\item{props}{properpensity score list containing the model}

\item{progs}{prognostic score list containing the model}

\item{treatment}{specify treatment}

\item{outcome}{specify outcome}

\item{covariates}{specify covariates}

\item{use_denom}{which denominator to use}

\item{joint_vip}{plot to add to}

\item{use_abs}{whether using absolute value or not}

\item{post_prop}{post analysis propensity score}

\item{post_prog}{post analysis prognostic score}
}
\description{
Get standardized difference in analysis sample
}
