% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/get_post_measures.R
=======
% Please edit documentation in R/measures.R
>>>>>>> 38802c698b7513bb5bd529d7c210b18f5081a4d5
\name{get_post_measures}
\alias{get_post_measures}
\title{Post-measures data frame to plot post-standardized omitted variable bias}
\usage{
get_post_measures(object, smd = "cross-sample")
}
\arguments{
\item{object}{post_jointVIP object}

\item{smd}{calculate standardized mean difference either using \code{cross-sample} or \code{pooled}}
}
\value{
measures needed for jointVIP
}
\description{
Post-measures data frame to plot post-standardized omitted variable bias
}
