% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_measures.R
\name{get_progs}
\alias{get_progs}
\title{Plot prognostic score histogram and get prognostic score.}
\usage{
get_progs(pilot_df, analysis_df, covariates, outcome)
}
\arguments{
\item{pilot_df}{dataframe of the pilot data}

\item{analysis_df}{dataframe of the analysis data}

\item{covariates}{vector of strings or list denoting column names of interest}

\item{outcome}{string denoting the name of the outcome variable}
}
\description{
Plot prognostic score histogram and get prognostic score.
}
