% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_measures.R
\name{get_props}
\alias{get_props}
\title{Plot propensity score comparison between treatment and control
and get propensity score.}
\usage{
get_props(pilot_df, analysis_df, covariates, treatment)
}
\arguments{
\item{pilot_df}{dataframe of the pilot data}

\item{analysis_df}{dataframe of the analysis data}

\item{covariates}{vector of strings or list denoting column names of interest}

\item{treatment}{string denoting the name of the treatment variable}
}
\description{
Plot propensity score comparison between treatment and control
and get propensity score.
}
