% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in R/support.R
=======
% Please edit documentation in R/general.R
>>>>>>> 38802c698b7513bb5bd529d7c210b18f5081a4d5
=======
% Please edit documentation in R/general.R
>>>>>>> 38802c698b7513bb5bd529d7c210b18f5081a4d5
\name{one_hot}
\alias{one_hot}
\title{support function for one-hot encoding}
\usage{
one_hot(df)
}
\arguments{
\item{df}{data.frame object for performing one-hot encoding}
}
\value{
data.frame object with factor variables one-hot encoded for each level
}
\description{
support function for one-hot encoding
}
