% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plot.jointVIP}
\alias{plot.jointVIP}
\title{plot the jointVIP object}
\usage{
\method{plot}{jointVIP}(
  x,
  ...,
  smd = "OVB-based",
  use_abs = TRUE,
  plot_title = "Joint Variable Importance Plot"
)
}
\arguments{
\item{x}{a jointVIP object}

\item{...}{custom options: \code{bias_curve_cutoffs}, \code{text_size}, \code{max.overlaps}, \code{label_cut_std_md}, \code{label_cut_outcome_cor}, \code{label_cut_bias}, \code{OVB_curves}, \code{add_var_labs}, \code{expanded_y_curvelab}}

\item{smd}{specify the standardized mean difference is \code{OVB-based} or \code{standard}}

\item{use_abs}{TRUE (default) for absolute measures}

\item{plot_title}{optional string for plot title}
}
\value{
a joint variable importance plot of class \code{ggplot}
}
\description{
plot the jointVIP object
}
\examples{
# not run
# new_jointVIP = create_jointVIP(treatment, outcome,
#                                covariates,
#                                pilot_df, analysis_df)

# plot(new_jointVIP, smd='OVB-based', use_abs=TRUE)
}
