% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_boot_variation.R
\name{plot_bootstrap}
\alias{plot_bootstrap}
\title{Wrapper for bootstrap to plot}
\usage{
plot_bootstrap(og_measures, boot_sd, use_denom, joint_vip, use_abs)
}
\arguments{
\item{og_measures}{original measures}

\item{boot_sd}{from bootstrap sd}

\item{use_denom}{which denom to use}

\item{joint_vip}{plot to add on to}

\item{use_abs}{whether to use absolute values or not}
}
\description{
Wrapper for bootstrap to plot
}
