% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jointVIP.R
\name{plot_jointVIP}
\alias{plot_jointVIP}
\title{Wrapper for the plot}
\usage{
plot_jointVIP(
  df,
  treatment,
  covariates,
  outcome,
  pilot_prop = 0.2,
  seed = 2521323,
  bias_curve_cutoffs = NULL,
  use_abs = F,
  use_pilot_denom = F,
  plot_title = "Joint variable importance"
)
}
\arguments{
\item{df}{data.frame of the original data}

\item{treatment}{string denoting the name of the treatment variable}

\item{covariates}{vector of strings or list denoting column names of interest}

\item{outcome}{string denoting the name of the outcome variable}

\item{pilot_prop}{proportion denoted as the pilot sample}

\item{seed}{int random seed}

\item{bias_curve_cutoffs}{cut offs of the bias curve one wishes to plot}

\item{use_abs}{boolean to denote whether absolute values are used to construct the measures}

\item{use_pilot_denom}{if true uses the pilot denominator}

\item{plot_title}{title of the plot to plot}
}
\value{
a list
}
\description{
Wrapper for the plot
}
