% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{print.post_jointVIP}
\alias{print.post_jointVIP}
\title{Obtains a print for post_jointVIP object}
\usage{
\method{print}{post_jointVIP}(x, ..., smd = "OVB-based", use_abs = TRUE, bias_tol = 0.01)
}
\arguments{
\item{x}{a post_jointVIP object}

\item{...}{not used}

\item{smd}{specify the standardized mean difference is \code{OVB-based} or \code{standard}}

\item{use_abs}{TRUE (default) for absolute measures}

\item{bias_tol}{numeric 0.01 (default) any bias above the absolute bias_tol will be printed}
}
\value{
measures used to create the plot of jointVIP
}
\description{
Obtains a print for post_jointVIP object
}
