% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{summary.jointVIP}
\alias{summary.jointVIP}
\title{Obtains a summary jointVIP object}
\usage{
\method{summary}{jointVIP}(object, ..., smd = "OVB-based", use_abs = TRUE, bias_tol = 0.01)
}
\arguments{
\item{object}{a jointVIP object}

\item{...}{not used}

\item{smd}{specify the standardized mean difference is \code{OVB-based} or \code{standard}}

\item{use_abs}{TRUE (default) for absolute measures}

\item{bias_tol}{numeric 0.01 (default) any bias above the absolute bias_tol will be summarized}
}
\description{
Obtains a summary jointVIP object
}
