% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{summary.post_jointVIP}
\alias{summary.post_jointVIP}
\title{Obtains a summary post_jointVIP object}
\usage{
\method{summary}{post_jointVIP}(
  object,
  ...,
  smd = "OVB-based",
  use_abs = TRUE,
  bias_tol = 0.01,
  post_bias_tol = 0.005
)
}
\arguments{
\item{object}{a post_jointVIP object}

\item{...}{not used}

\item{smd}{specify the standardized mean difference is \code{OVB-based} or \code{standard}}

\item{use_abs}{TRUE (default) for absolute measures}

\item{bias_tol}{numeric 0.01 (default) any bias above the absolute bias_tol will be summarized}

\item{post_bias_tol}{numeric 0.005 (default) any bias above the absolute bias_tol will be summarized}
}
\description{
Obtains a summary post_jointVIP object
}
