% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{add_bias_curves}
\alias{add_bias_curves}
\title{support function to plot bias curves}
\usage{
add_bias_curves(p, ...)
}
\arguments{
\item{p}{plot made with jointVIP object}

\item{...}{encompasses other variables needed}
}
\value{
a joint variable importance plot of class \code{ggplot} with curves
}
\description{
support function to plot bias curves
}
