% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in R/support.R
=======
% Please edit documentation in R/general.R
>>>>>>> 38802c698b7513bb5bd529d7c210b18f5081a4d5
=======
% Please edit documentation in R/general.R
>>>>>>> 38802c698b7513bb5bd529d7c210b18f5081a4d5
\name{ceiling_dec}
\alias{ceiling_dec}
\title{support function for ceiling function with decimals}
\usage{
ceiling_dec(num, dec_place = 1)
}
\arguments{
\item{num}{numeric}

\item{dec_place}{decimal place that is desired ceiling for}
}
\value{
numeric number desired
}
\description{
support function for ceiling function with decimals
}
