% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ceiling_dec}
\alias{ceiling_dec}
\title{support function for ceiling function with decimals}
\usage{
ceiling_dec(num, dec_place = 1)
}
\arguments{
\item{num}{numeric}

\item{dec_place}{decimal place that is desired ceiling for}
}
\value{
numeric number desired
}
\description{
support function for ceiling function with decimals
}
