% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/support.R
=======
% Please edit documentation in R/general.R
>>>>>>> 38802c698b7513bb5bd529d7c210b18f5081a4d5
\name{floor_dec}
\alias{floor_dec}
\title{support function for floor function with decimals}
\usage{
floor_dec(num, dec_place = 1)
}
\arguments{
\item{num}{numeric}

\item{dec_place}{decimal place that is desired floor for}
}
\value{
numeric number desired
}
\description{
support function for floor function with decimals
}
