% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{floor_dec}
\alias{floor_dec}
\title{support function for floor function with decimals}
\usage{
floor_dec(num, dec_place = 1)
}
\arguments{
\item{num}{numeric}

\item{dec_place}{decimal place that is desired floor for}
}
\value{
numeric number desired
}
\description{
support function for floor function with decimals
}
