% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jointVIP.R
\name{get_jointVIP}
\alias{get_jointVIP}
\title{Get the measures and create plot}
\usage{
get_jointVIP(
  measures,
  use_pilot_denom = F,
  bias_curve_cutoffs = c(-0.05, -0.03, -0.01, -0.005, 0.005, 0.01, 0.03, 0.05),
  use_abs = F,
  plot_title = "Joint variable importance"
)
}
\arguments{
\item{measures}{the measures calculated for jointVIP}

\item{use_pilot_denom}{if true uses the pilot denominator}

\item{bias_curve_cutoffs}{cut offs of the bias curve one wishes to plot}

\item{use_abs}{boolean to denote whether absolute values are used to construct the measures}

\item{plot_title}{title of the plot to plot}
}
\description{
Get the measures and create plot
}
