% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_measure.R
\name{get_measures}
\alias{get_measures}
\title{Create a data frame filled with the difference and denominator
of both the analysis sample and pilot sample.
It also contains pilot vs outcome correlation.
This is done for all the covariates.}
\usage{
get_measures(
  df,
  covariates,
  treatment,
  outcome,
  pilot_prop = 0.2,
  seed = 2521323,
  use_abs = F
)
}
\arguments{
\item{df}{data.frame of the original data}

\item{covariates}{vector of strings or list denoting column names of interest}

\item{treatment}{string denoting the name of the treatment variable}

\item{outcome}{string denoting the name of the outcome variable}

\item{pilot_prop}{proportion denoted as the pilot sample}

\item{seed}{random number to control for randomness, make it reproducible}

\item{use_abs}{boolean to denote whether absolute values are used to construct the measures}
}
\description{
Create a data frame filled with the difference and denominator
of both the analysis sample and pilot sample.
It also contains pilot vs outcome correlation.
This is done for all the covariates.
}
