% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_measures.R
\name{get_measures}
\alias{get_measures}
\title{Create a data frame filled with the difference and denominator
of both the analysis sample and pilot sample.
It also contains pilot vs outcome correlation.
This is done for all the covariates.}
\usage{
get_measures(
  pilot_df,
  analysis_df,
  covariates,
  treatment,
  outcome,
  use_abs = F,
  user_props = NULL,
  user_progs = NULL,
  bias_tol = 0.005
)
}
\arguments{
\item{pilot_df}{dataframe of the pilot data}

\item{analysis_df}{dataframe of the analysis data}

\item{covariates}{vector of strings or list denoting column names of interest}

\item{treatment}{string denoting the name of the treatment variable}

\item{outcome}{string denoting the name of the outcome variable}

\item{use_abs}{boolean to denote whether absolute values are used to construct the measures}

\item{user_props}{user input propensity score for pilot and analysis sample, and model fit; if null calculates logistic regression}

\item{user_progs}{user input prognostic score for pilot and analysis sample, and model fit; if null calculates logistic/ linear regression}

\item{bias_tol}{bias tolerance; note always calculated by pilot sample}
}
\description{
Create a data frame filled with the difference and denominator
of both the analysis sample and pilot sample.
It also contains pilot vs outcome correlation.
This is done for all the covariates.
}
