% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{get_measures}
\alias{get_measures}
\title{Prepare data frame to plot standardized omitted variable bias
Marginal standardized mean differences and outcome correlation}
\usage{
get_measures(object, smd = "cross-sample")
}
\arguments{
\item{object}{jointVIP object}

\item{smd}{calculate standardized mean difference either using \code{cross-sample} or \code{pooled}}
}
\value{
measures needed for jointVIP
}
\description{
Prepare data frame to plot standardized omitted variable bias
Marginal standardized mean differences and outcome correlation
}
