% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_measures.R
\name{get_pilot_cor}
\alias{get_pilot_cor}
\title{Get correlation in pilot dataframe}
\usage{
get_pilot_cor(covariate, pilot_df, outcome)
}
\arguments{
\item{covariate}{covariate, note it is each column of the sample}

\item{pilot_df}{dataframe of the pilot data}

\item{outcome}{string denoting the name of the outcome variable}
}
\description{
Get correlation in pilot dataframe
}
