% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_measure.R
\name{get_progs}
\alias{get_progs}
\title{Plot prognostic score histogram and get prognostic score.}
\usage{
get_progs(df, covariates, outcome, pilot_sample_num)
}
\arguments{
\item{df}{data.frame of the original data}

\item{covariates}{vector of strings or list denoting column names of interest}

\item{outcome}{string denoting the name of the outcome variable}

\item{pilot_sample_num}{vector of numbers indicating which samples are denoted as the pilot sample}
}
\description{
Plot prognostic score histogram and get prognostic score.
}
