% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_measure.R
\name{get_props}
\alias{get_props}
\title{Plot propensity score comparison between treatment and control
and get propensity score.}
\usage{
get_props(df, covariates, treatment, pilot_sample_num)
}
\arguments{
\item{df}{data.frame of the original data}

\item{covariates}{vector of strings or list denoting column names of interest}

\item{treatment}{string denoting the name of the treatment variable}

\item{pilot_sample_num}{vector of numbers indicating which samples are denoted as the pilot sample}
}
\description{
Plot propensity score comparison between treatment and control
and get propensity score.
}
