% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{one_hot}
\alias{one_hot}
\title{support function for one-hot encoding}
\usage{
one_hot(df)
}
\arguments{
\item{df}{data.frame object for performing one-hot encoding}
}
\value{
data.frame object with factor variables one-hot encoded for each level
}
\description{
support function for one-hot encoding
}
