% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jointVIP.R
\name{plot_jointVIP}
\alias{plot_jointVIP}
\title{Wrapper for the plot}
\usage{
plot_jointVIP(
  pilot_df,
  analysis_df,
  treatment,
  covariates,
  outcome,
  raw_plot = F,
  props = NULL,
  progs = NULL,
  post_analysis_df = NULL,
  post_prop = NULL,
  post_prog = NULL,
  bias_curve_cutoffs = NULL,
  use_abs = F,
  use_denom = "both",
  plot_title = "Joint variable importance",
  label_cutoff_std_diff = NULL,
  label_cutoff_control_cor = NULL,
  label_cutoff_bias = NULL
)
}
\arguments{
\item{pilot_df}{dataframe of the pilot data}

\item{analysis_df}{dataframe of the analysis data}

\item{treatment}{string denoting the name of the treatment variable}

\item{covariates}{vector of strings or list denoting column names of interest}

\item{outcome}{string denoting the name of the outcome variable}

\item{raw_plot}{plot without bias}

\item{props}{user input propensity score list}

\item{progs}{user input prognostic score list}

\item{post_analysis_df}{post match analysis data frame}

\item{post_prop}{user input propensity score only used when post_analysis_df is used}

\item{post_prog}{user input prognostic score only used when post_analysis_df is used}

\item{bias_curve_cutoffs}{cut offs of the bias curve one wishes to plot}

\item{use_abs}{boolean to denote whether absolute values are used to construct the measures}

\item{use_denom}{must be "standard", "pilot", or "min"}

\item{plot_title}{title of the plot to plot}

\item{label_cutoff_std_diff}{text label cut off for standardized difference}

\item{label_cutoff_control_cor}{text label cut off for correlation}

\item{label_cutoff_bias}{text label cut off for bias}
}
\value{
a list
}
\description{
Wrapper for the plot
}
